Use ES rest template over repository

+++ ontopop-search/src/main/java/ai/hyperlearning/ontopop/search/elasticsearch/ElasticsearchService.java
@@ -1,10 +1,19 @@
 package ai.hyperlearning.ontopop.search.elasticsearch;
 
+import static org.elasticsearch.index.query.QueryBuilders.fuzzyQuery;
+import static org.elasticsearch.index.query.QueryBuilders.matchQuery;
+
 import java.util.Set;
 
+import org.elasticsearch.index.query.Operator;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
+import org.springframework.data.elasticsearch.core.SearchHits;
 import org.springframework.data.elasticsearch.core.mapping.IndexCoordinates;
+import org.springframework.data.elasticsearch.core.query.NativeSearchQueryBuilder;
+import org.springframework.data.elasticsearch.core.query.Query;
 import org.springframework.stereotype.Service;
 
 import ai.hyperlearning.ontopop.search.SearchService;
@@ -20,6 +29,9 @@ import ai.hyperlearning.ontopop.search.model.SimpleIndexedVertex;
 @Service
 public class ElasticsearchService implements SearchService {
 	
+	private static final Logger LOGGER = 
+			LoggerFactory.getLogger(ElasticsearchService.class);
+	
 	@Autowired
 	private ElasticsearchOperations elasticsearchTemplate;
 	
@@ -29,20 +41,75 @@ public class ElasticsearchService implements SearchService {
 
 	@Override
 	public void createIndex(String indexName) {
-		elasticsearchTemplate.indexOps(
-				IndexCoordinates.of(indexName)).create();
+		if (!elasticsearchTemplate.indexOps(
+				IndexCoordinates.of(indexName)).exists()) {
+			elasticsearchTemplate.indexOps(
+					IndexCoordinates.of(indexName)).create();
+		} else {
+			LOGGER.warn("Index '{}' already exists. Skipping index creation.",
+					indexName);
+		}
 	}
 	
 	@Override
 	public void deleteIndex(String indexName) {
-		elasticsearchTemplate.indexOps(
-				IndexCoordinates.of(indexName)).delete();
+		if (elasticsearchTemplate.indexOps(
+				IndexCoordinates.of(indexName)).exists()) {
+			elasticsearchTemplate.indexOps(
+					IndexCoordinates.of(indexName)).delete();
+		} else {
+			LOGGER.warn("Index '{}' does not exist. Skipping index deletion.", 
+					indexName);
+		}
 	}
 	
 	/**************************************************************************
 	 * DOCUMENT MANAGEMENT
 	 *************************************************************************/
 
+	@Override
+	public SimpleIndexedVertex getDocument(String indexName, long vertexId) {
+		return elasticsearchTemplate.get(String.valueOf(vertexId), 
+				SimpleIndexedVertex.class, 
+				IndexCoordinates.of(indexName));
+	}
+	
+	@Override
+	public SearchHits<SimpleIndexedVertex> search(String indexName, 
+			String propertyKey, String query, boolean exact, boolean and) {
+		
+		Operator operator = and ? Operator.AND : Operator.OR;
+		final Query searchQuery = exact ? new NativeSearchQueryBuilder()
+				.withQuery(matchQuery(propertyKey, query)
+						.operator(operator))
+	            .build() : new NativeSearchQueryBuilder()
+					.withQuery(fuzzyQuery(propertyKey, query))
+					.build();
+		return elasticsearchTemplate
+				.search(searchQuery, SimpleIndexedVertex.class, 
+						IndexCoordinates.of(indexName));
+		
+	}
+	
+	@Override
+	public SearchHits<SimpleIndexedVertex> search(String indexName, 
+			String propertyKey, String query, boolean and, 
+			int minimumShouldMatchPercentage) {
+		
+		Operator operator = and ? Operator.AND : Operator.OR;
+		String minimumShouldMatchPct = String.valueOf(
+				minimumShouldMatchPercentage) + "%";
+		final Query searchQuery = new NativeSearchQueryBuilder()
+				.withQuery(matchQuery(propertyKey, query)
+						.operator(operator)
+						.minimumShouldMatch(minimumShouldMatchPct))
+	            .build();
+		return elasticsearchTemplate
+				.search(searchQuery, SimpleIndexedVertex.class, 
+						IndexCoordinates.of(indexName));
+		
+	}
+	
 	@Override
 	public void indexDocuments(String indexName, 
 			Set<SimpleIndexedVertex> vertices) {
